requirements.txt
pillow
requests
pysocks
flask
flask-cors
pyyaml